feat: Add route for home page and set default view engine
The code now includes a route for the home page that displays index.html when someone enters the URL https://sensgro.rec22test.site/. Additionally, the default view engine is now set to EJS, which allows for easier rendering of dynamic templates.

This commit also updates the error message for when no default engine is specified when rendering templates.